requirements.txt
flask==3.0.0
ultralytics==8.0.196
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3